# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  domain.SellerAttributes:
    properties:
      address:
        type: string
      cid:
        type: string
      company_name:
        type: string
      locality_id:
        type: string
      telephone:
        type: string
    type: object
  domain.SellerDoc:
    properties:
      address:
        type: string
      cid:
        type: string
      company_name:
        type: string
      id:
        type: integer
      locality_id:
        type: string
      telephone:
        type: string
    type: object
  domain.SellerResponse:
    properties:
      data:
        $ref: '#/definitions/domain.SellerDoc'
    type: object
  domain.SellersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.SellerDoc'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestionar sellers y warehouses
  title: API de Sellers y Warehouses
  version: "1.0"
paths:
...
  /sellers:
    get:
      description: Retorna una lista de todos los proveedores registrados en el sistema
      produces:
      - application/json
      responses:
        "200":
          description: Listado de proveedores
          schema:
            $ref: '#/definitions/domain.SellersResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Lista todos los proveedores
      tags:
      - Sellers
    post:
      consumes:
      - application/json
      description: Crea un nuevo proveedor en el sistema con los datos proporcionados
      parameters:
      - description: Datos del proveedor a crear
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/domain.SellerAttributes'
      produces:
      - application/json
      responses:
        "201":
          description: Proveedor creado exitosamente
          schema:
            $ref: '#/definitions/domain.SellerResponse'
        "409":
          description: El CID ya existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Datos malformados o faltantes
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Crea un nuevo proveedor
      tags:
      - Sellers
  /sellers/{id}:
    delete:
      description: Elimina un proveedor existente de la base de datos
      parameters:
      - description: ID del proveedor
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Proveedor eliminado exitosamente
        "400":
          description: ID no enviado o malformado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: El proveedor no existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Elimina un proveedor por su ID
      tags:
      - Sellers
    get:
      description: Retorna la información de un proveedor específico basado en su
        ID
      parameters:
      - description: ID del proveedor
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Proveedor encontrado
          schema:
            $ref: '#/definitions/domain.SellerResponse'
        "400":
          description: ID no enviado o malformado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: El proveedor no existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Retorna un proveedor por su ID
      tags:
      - Sellers
    patch:
      consumes:
      - application/json
      description: Actualiza un proveedor existente con los datos proporcionados
      parameters:
      - description: ID del proveedor
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del proveedor a actualizar
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/domain.SellerAttributes'
      produces:
      - application/json
      responses:
        "200":
          description: Proveedor actualizado exitosamente
          schema:
            $ref: '#/definitions/domain.SellerResponse'
        "400":
          description: ID no enviado o malformado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: El proveedor no existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: El CID ya existe
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Datos malformados o faltantes
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Actualiza un proveedor por su ID
      tags:
      - Sellers
...
swagger: "2.0"